Analyst for Overall Winners
Did a bit to see if there's any resemblance or variables that we should use instead of all.
Moving forward, will need to find a reason as to why we are using depth level 20 and the variables. As of now, the classification accuracy is almost the same for all variables which I will have to look into again.

Moving forward:
- Analysis of which AI classification model is best for our cast
- Take a look at the analyst done by martin with the boxplot etc
- Compile all analysis into one notebook